backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg
aiosqlite==0.19.0
psycopg2-binary==2.9.9
langchain
openai==1.3.5
slack-sdk==3.26.1
apscheduler==3.10.4
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
requests==2.31.0
tzlocal==5.2
pytz==2023.3
alembic==1.12.1 